Extract useCollectionList and CollectionSortMenu
